Fixed a bug - the former code used listProp where it should have used the list in the Prop instead. This lead to bad change propagation.
The code now retains (where possible, and invalidates as necessary) a weak reference to the current list. This list is then used to check changes. If the indexProp has no changes, the listProp has only changes with sameInstances, and the current list is known and has only sameInstances changes, then the Change is propagated as sameInstances. Otherwise, it is propagated as changing instances.