Possible fix for issue where "fragile" listeners have problems when receiving events. This seems to occur when multiple table models in the composite change "at the same time", in the sense that the underlying data of two models change, but one obviously gets the chance to fire an event to the composite table model first.
Say for example we have two table models A and B, in a composite C. A and B both watch the same underlying data. The underlying data changes in such a way that A has some small change, for example to a few rows, but not the structure of the data - if the data is queried, all columns and rows will be valid. B also changes, but since it views the data in a different way it goes from having 2 columns to having only a single column.

If A "notices" that its data has changed first (perhaps it is notified via an event first, because it registered with the data first), then it will fire an event to the composite table model C first. C assesses the event, finds that it does NOT affect columns, and so fires on an event saying that the structure of C has NOT changed. It does not know about the event "in the queue" that it will receive from B, saying that B's columns HAVE changed. Therefore the event fired by C is too small, and fragile listeners may therefore have index out of bounds exceptions - this happens with sorted tables, but apparently not with unsorted tables.

Of course, C will soon fire another event saying that the columns HAVE changed, when it gets the event from B, but it is now too late for the fragile listener.

To fix this, C now stores its previous row AND column sizes (used to just do rows). Hence we can actually update sizes and interrogate all models directly to check their new sizes BEFORE firing any event, so that the even is always large enough to "protect" fragile listeners. This stops them assuming that columns are still present when they may have disappeared, for example.

We still respect incoming event sizes - even if it seems that the column count hasn't changed, we fire a structure change regardless, since the change may be to the arrangement of columns rather than count, etc. So the worst case is now that we say columns haven't changed when in fact they have changed class/name etc., but we know that there are still enough columns present. It is hard to do better than this without somehow caching events from all models until we know we have  a "set", but it is not clear how we would know we have a consistent set of changes, unless we could query models to see if they have pending changes, which would actually be a nice extension to event based changes.