Fixed quite subtle bug - JViews aren't kept alive by the data they are listening to, and generally aren't very interesting to their users - when making a UI it is common to just make some JViews, put their JComponents into the UI, then forget about the JViews. Then the JViews are GCed, leaving the orphan JComponents without updates and other logic.
To fix this, each JView implementation now makes sure that the JComponent it hands out via getComponent() also has a link back to the JView itself. Hence the JView will remain in scope as long as its JComponent does. Of course once the JComponent is no longer in scope, the JView will be GCed as required (unless there is a separate strong link to it - the Box->View link is weak and so allows the JView to be GCed). This is very useful since we don't need to worry about having a "dispose()" method, or removing Views from Boxes, the garbage collector will get rid of things as necessary, for us.