Use a new "RetainingPanel" as the component of this view, which wraps the actual component, but also serves to retain the view to avoid it being GCed. This allows for the common case where a user of a View just retains the component it applies (in their tree of Swing components) and loses the reference to the View itself, leaving the View open to being collected.
This should make no difference to users - changes to background and foreground colours are passed through to the contained component, and the panel is not opaque. However there might be minor glitches, for example in the case of attempting to set component font etc., which can be fixed as they are encountered.