Removed PainTer println.
RotatedTableCellRenderer delegates to an existing TableCellRenderer (most likely the default one from the table we plan to use the rotated header with). Text is cleared, and rendered as an icon with vertical alignment. This is a workaround to allow inheriting mysterious underlying styling, etc.

Renderer for booleans, displays colored rects, red for true, transparent for false - useful for error states.